requirements.txt
ipdb==0.8.1
marathon==0.7.6
PyStaticConfiguration==0.9.0